refactor: ErrorMessages 클래스를 configure 패키지로 이동
