+++ src/main/java/lotto/configure/ErrorMessageHelper.java
@@ -1,5 +1,5 @@
 package lotto.configure;
 
-public class ErrorMessage {
+public class ErrorMessageHelper {
     public static final String ERROR_MESSAGE_COMMON_PREFIX = "[ERROR]";
 }
